Move license server type to constants module
