+++ agent/lm_agent/backend_utils/constants.py
@@ -1,0 +1,13 @@
+from enum import Enum
+
+
+class LicenseServerType(str, Enum):
+    """
+    Describe the supported license server types that may be used for fetching licenses from license servers.
+    """
+
+    FLEXLM = "flexlm"
+    RLM = "rlm"
+    LMX = "lmx"
+    LSDYNA = "lsdyna"
+    OLICENSE = "olicense"
